requirements.txt
streamlit==1.33.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0
